requirements.txt
Flask==3.0.3
waitress==3.0.0
gpiozero==2.0.1
python-dotenv==1.0.1
pytest==8.2.0
